requirements.txt
yfinance
pandas
genai
load_dotenv
reportlab
requests
beautifulsoup4
